misc/HtmlTestSuite/results/test-0440.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>Table quirks</title>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/resources/check-layout-th.js"/>
<meta content="ahem" name="flags"/>
<link href="/fonts/ahem.css" rel="stylesheet" type="text/css"/>
<link href="./support/table-tentative.css" rel="stylesheet" type="text/css"/>
<link href="atotic@chromium.org" rel="author" title="Aleks Totic"/>
<link href="https://quirks.spec.whatwg.org/#the-table-cell-width-calculation-quirk" rel="help"/>
<style>
 table {
    table-layout: auto;
    border-spacing: 8px 8px;
  }
  td {
    padding: 20px;
    background: #BFB;
    font-size: 10px;
    box-sizing: border-box;
  }
  td &gt; div {
    display: inline-block;
    background: rgba(56,162,56,0.3);
  }
</style>

</head><body><h1>Tables in quirks mode proposals</h1>

<p><a href="https://quirks.spec.whatwg.org/#the-table-cell-height-box-sizing-quirk">The table cell height box sizing quirk</a></p>
<table data-expected-height="116">
  <tbody><tr>
    <td data-expected-height="100" style="height:100px;box-sizing:content-box">100 height</td>
  </tr>
</tbody></table>
<table data-expected-height="116">
  <tbody><tr>
    <td data-expected-height="100" style="height:100px;box-sizing:border-box">100 height</td>
  </tr>
</tbody></table>

<p><a href="https://quirks.spec.whatwg.org/#the-table-cell-nowrap-minimum-width-calculation-quirk">The table cell nowrap minimum width calculation quirk</a></p>
<p class="error">Chrome Legacy, Edge, Safari fail, only FF gets it correct.<b>Proposal: deprecate the quirk</b></p>
<table>
  <tbody><tr>
    <td data-expected-width="580" nowrap="" style="width:50px;font: 20px/1 Ahem">nowrap nowrap nowrap nowrap</td>
  </tr>
</tbody></table>

<p><a href="https://quirks.spec.whatwg.org/#the-text-decoration-doesnt-propagate-into-tables-quirk">The text decoration doesn’t propagate into tables quirk</a></p>
<div style="font-style:italic">
  <table>
    <tbody><tr><td id="notitalic">decoration</td>
  </tr></tbody></table>
</div>

<p><a href="https://quirks.spec.whatwg.org/#the-collapsing-table-quirk">The collapsing table quirk</a></p>
<p class="error">Chrome Legacy/Edge/Safari ignore the quirk, FF does not. <b>Proposal: depreciate the quirk</b></p>
<table data-expected-height="40" data-expected-width="40" style="border: 20px solid green"/>

<p><a href="https://quirks.spec.whatwg.org/#the-table-cell-width-calculation-quirk">The table cell width calculation quirk</a></p>
<table style="width:200px">
  <tbody><tr><td data-expected-width="290"><img style="width:50px;height:20px"/><img style="width:50px;height:20px"/><img style="width:50px;height:20px"/><img style="width:50px;height:20px"/><img style="width:50px;height:20px"/></td>
</tr></tbody></table>

<script>
  let pngSrc="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAMgAAACWAQMAAAChElVaAAAABGdBTUEAALGPC/xhBQAAACBjSFJNAAB6JgAAgIQAAPoAAACA6AAAdTAAAOpgAAA6mAAAF3CculE8AAAABlBMVEUAgAD///8UPy9PAAAAAWJLR0QB/wIt3gAAAAd0SU1FB+MBDwkdA1Cz/EMAAAAbSURBVEjH7cGBAAAAAMOg+VPf4ARVAQAAAM8ADzwAAeM8wQsAAAAldEVYdGRhdGU6Y3JlYXRlADIwMTktMDEtMTVUMTc6Mjk6MDMtMDg6MDCYDy9IAAAAJXRFWHRkYXRlOm1vZGlmeQAyMDE5LTAxLTE1VDE3OjI5OjAzLTA4OjAw6VKX9AAAAABJRU5ErkJggg=="
;
  for (let img of Array.from(document.querySelectorAll("img"))) {
    img.src = pngSrc;
  }
  test(_ =&gt; {
    assert_equals(window.getComputedStyle(document.querySelector("#notitalic")).fontStyle, "normal");
  }, "decoration does not propagate into table");
  checkLayout("table");
</script>
</body></html>
